Updates Japan to now work with CODA 3 (Coda 2 compatibility is broken). Made several functions in CodaDecoder.h virtual and overrode in QwEventBuffer.h as well as overrode pre-exisiting virtual functions.
Main changes:
CodaDecoder::LoadEvent -> QwEventBuffer::DecodeEvent
Adjusted where key variables were defined to be in accordance with the new CODA 3 data format. There are two parallel variables (one is used in Podd and the other is used internally in JAPAN) of the variables below:
   * fCurrentRun
   * fEvtType
   * fEvtNumber
   * fBankDataType
	 * fWordsSoFar
TODO: We will need to decide if we should keep both or choose one to use everywhere for consistency.

Removed a lot of Podd-specific Error checking (ie all mention of the THaCrateMap.h and its accompanying checks).

Regressions Introduced:
* Coda 2 format no longer works
* User events no longer work
* Coda version hardcoded in
* Reduced variable consistency